+++ example-extensions/src/main/java/com/zhuinden/simplestackextensionsample/app/AuthenticationManager.kt
@@ -1,26 +1,35 @@
 package com.zhuinden.simplestackextensionsample.app
 
-import android.content.Context
-import android.preference.PreferenceManager
+import android.content.SharedPreferences
 
-object AuthenticationManager {
-    @Suppress("DEPRECATION") // w/e androidx-chan
-    fun isAuthenticated(appContext: Context): Boolean {
-        val sharedPref = PreferenceManager.getDefaultSharedPreferences(appContext)
-        return sharedPref.getBoolean("isRegistered", false)
+class AuthenticationManager(
+    private val sharedPreferences: SharedPreferences
+) {
+    companion object {
+        private const val KEY_IS_REGISTERED = "isRegistered"
+        private const val KEY_REGISTERED_USERNAME = "registeredUsername"
     }
 
-    @Suppress("DEPRECATION") // w/e androidx-chan
-    fun saveRegistration(appContext: Context) {
-        val sharedPref = PreferenceManager.getDefaultSharedPreferences(appContext)
-        sharedPref.edit().putBoolean("isRegistered", true).apply()
+    fun isAuthenticated(): Boolean =
+        sharedPreferences.getBoolean(KEY_IS_REGISTERED, false)
+
+    fun saveRegistration(username: String) {
+        sharedPreferences.edit()
+            .putBoolean(KEY_IS_REGISTERED, true)
+            .putString(KEY_REGISTERED_USERNAME, username)
+            .apply()
     }
 
-    @Suppress("DEPRECATION") // w/e androidx-chan
-    fun clearRegistration(appContext: Context) {
-        val sharedPref = PreferenceManager.getDefaultSharedPreferences(appContext)
-        sharedPref.edit().remove("isRegistered").apply()
+    fun clearRegistration() {
+        sharedPreferences.edit()
+            .remove(KEY_IS_REGISTERED)
+            .remove(KEY_REGISTERED_USERNAME)
+            .apply()
     }
 
-    var authToken: String = "" // why would this be in the viewModel?
+    fun getAuthenticatedUser(): String {
+        val username = sharedPreferences.getString(KEY_REGISTERED_USERNAME, "").takeUnless { it.isNullOrEmpty() }
+
+        return checkNotNull(username)
+    }
 }