+++ example-extensions/src/main/java/com/zhuinden/simplestackextensionsample/features/login/LoginViewModel.kt
@@ -1,21 +1,20 @@
 package com.zhuinden.simplestackextensionsample.features.login
 
-import android.content.Context
 import com.jakewharton.rxrelay2.BehaviorRelay
-import com.zhuinden.rxcombinetuplekt.combineTuple
+import com.zhuinden.rxvalidatebykt.validateBy
 import com.zhuinden.simplestack.*
 import com.zhuinden.simplestackextensionsample.app.AuthenticationManager
 import com.zhuinden.simplestackextensionsample.features.profile.ProfileKey
 import com.zhuinden.simplestackextensionsample.features.registration.EnterProfileDataKey
 import com.zhuinden.simplestackextensionsample.utils.get
+import com.zhuinden.simplestackextensionsample.utils.observe
 import com.zhuinden.simplestackextensionsample.utils.set
 import com.zhuinden.statebundle.StateBundle
+import io.reactivex.Observable
 import io.reactivex.disposables.CompositeDisposable
-import io.reactivex.rxkotlin.addTo
-import io.reactivex.rxkotlin.subscribeBy
 
 class LoginViewModel(
-    private val appContext: Context,
+    private val authenticationManager: AuthenticationManager,
     private val backstack: Backstack
 ) : Bundleable, ScopedServices.Registered {
     private val compositeDisposable = CompositeDisposable()
@@ -23,13 +22,16 @@ class LoginViewModel(
     val username = BehaviorRelay.createDefault("")
     val password = BehaviorRelay.createDefault("")
 
-    val isLoginEnabled = BehaviorRelay.createDefault(false)
+    private val isLoginEnabledRelay = BehaviorRelay.createDefault(false)
+    val isLoginEnabled: Observable<Boolean> = isLoginEnabledRelay
 
     override fun onServiceRegistered() {
-        combineTuple(username, password)
-            .subscribeBy { (username, password) ->
-                isLoginEnabled.set(username.isNotBlank() && password.isNotBlank())
-            }.addTo(compositeDisposable)
+        validateBy(
+            username.map { it.isNotBlank() },
+            password.map { it.isNotBlank() },
+        ).observe(compositeDisposable) {
+            isLoginEnabledRelay.set(it)
+        }
     }
 
     override fun onServiceUnregistered() {
@@ -37,14 +39,14 @@ class LoginViewModel(
     }
 
     fun onLoginClicked() {
-        if (isLoginEnabled.get()) {
-            AuthenticationManager.saveRegistration(appContext)
-            backstack.setHistory(History.of(ProfileKey()), StateChange.FORWARD)
+        if (isLoginEnabledRelay.get()) {
+            authenticationManager.saveRegistration(username.get())
+            backstack.setHistory(History.of(ProfileKey(username.get())), StateChange.FORWARD)
         }
     }
 
     fun onRegisterClicked() {
-        backstack.goTo(EnterProfileDataKey())
+        backstack.goTo(EnterProfileDataKey)
     }
 
     override fun toBundle(): StateBundle = StateBundle().apply {