+++ example-extensions/src/main/java/com/zhuinden/simplestackextensionsample/features/profile/ProfileViewModel.kt
@@ -1,6 +1,5 @@
 package com.zhuinden.simplestackextensionsample.features.profile
 
-import android.content.Context
 import com.zhuinden.simplestack.Backstack
 import com.zhuinden.simplestack.History
 import com.zhuinden.simplestack.ScopedServices
@@ -9,12 +8,12 @@ import com.zhuinden.simplestackextensionsample.app.AuthenticationManager
 import com.zhuinden.simplestackextensionsample.features.login.LoginKey
 
 class ProfileViewModel(
-    private val appContext: Context,
+    private val authenticationManager: AuthenticationManager,
     private val backstack: Backstack
 ) : ScopedServices.Activated {
     override fun onServiceActive() {
-        if (!AuthenticationManager.isAuthenticated(appContext)) {
-            backstack.setHistory(History.of(LoginKey()), StateChange.REPLACE)
+        if (!authenticationManager.isAuthenticated()) {
+            backstack.setHistory(History.of(LoginKey), StateChange.REPLACE)
         }
     }