+++ example-extensions/src/main/java/com/zhuinden/simplestackextensionsample/features/registration/CreateLoginCredentialsKey.kt
@@ -3,11 +3,11 @@ package com.zhuinden.simplestackextensionsample.features.registration
 import androidx.fragment.app.Fragment
 import com.zhuinden.simplestack.ScopeKey
 import com.zhuinden.simplestackextensions.fragments.DefaultFragmentKey
-import kotlinx.android.parcel.Parcelize
+import kotlinx.parcelize.Parcelize
 
 @Parcelize
-data class CreateLoginCredentialsKey(private val placeholder: String = "") : DefaultFragmentKey(), ScopeKey.Child {
+data object CreateLoginCredentialsKey : DefaultFragmentKey(), ScopeKey.Child {
     override fun instantiateFragment(): Fragment = CreateLoginCredentialsFragment()
 
-    override fun getParentScopes(): List<String> = listOf("registration")
+    override fun getParentScopes(): List<String> = listOf(RegistrationViewModel::class.java.name)
 }