+++ example-extensions/src/main/java/com/zhuinden/simplestackextensionsample/features/registration/EnterProfileDataFragment.kt
@@ -5,14 +5,9 @@ import android.view.View
 import com.zhuinden.simplestackextensions.fragments.KeyedFragment
 import com.zhuinden.simplestackextensions.fragmentsktx.lookup
 import com.zhuinden.simplestackextensionsample.R
-import com.zhuinden.simplestackextensionsample.utils.get
-import com.zhuinden.simplestackextensionsample.utils.onClick
-import com.zhuinden.simplestackextensionsample.utils.onTextChanged
-import com.zhuinden.simplestackextensionsample.utils.set
+import com.zhuinden.simplestackextensionsample.databinding.EnterProfileDataFragmentBinding
+import com.zhuinden.simplestackextensionsample.utils.*
 import io.reactivex.disposables.CompositeDisposable
-import io.reactivex.rxkotlin.addTo
-import io.reactivex.rxkotlin.subscribeBy
-import kotlinx.android.synthetic.main.enter_profile_data_fragment.*
 
 class EnterProfileDataFragment : KeyedFragment(R.layout.enter_profile_data_fragment) {
     private val viewModel by lazy { lookup<RegistrationViewModel>() }
@@ -22,16 +17,18 @@ class EnterProfileDataFragment : KeyedFragment(R.layout.enter_profile_data_fragm
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        textFullName.setText(viewModel.fullName.get())
-        textBio.setText(viewModel.bio.get())
+        val binding = EnterProfileDataFragmentBinding.bind(view)
 
-        viewModel.isEnterProfileNextEnabled.distinctUntilChanged().subscribeBy { enabled ->
-            buttonEnterProfileNext.isEnabled = enabled
-        }.addTo(compositeDisposable)
+        binding.textFullName.setText(viewModel.fullName.get())
+        binding.textBio.setText(viewModel.bio.get())
 
-        textFullName.onTextChanged { fullName -> viewModel.fullName.set(fullName) }
-        textBio.onTextChanged { bio -> viewModel.bio.set(bio) }
-        buttonEnterProfileNext.onClick { viewModel.onEnterProfileNextClicked() }
+        viewModel.isEnterProfileNextEnabled.observe(compositeDisposable) { enabled ->
+            binding.buttonEnterProfileNext.isEnabled = enabled
+        }
+
+        binding.textFullName.onTextChanged { fullName -> viewModel.fullName.set(fullName) }
+        binding.textBio.onTextChanged { bio -> viewModel.bio.set(bio) }
+        binding.buttonEnterProfileNext.onClick { viewModel.onEnterProfileNextClicked() }
     }
 
     override fun onDestroyView() {