+++ example-extensions/src/main/java/com/zhuinden/simplestackextensionsample/features/registration/EnterProfileDataKey.kt
@@ -3,11 +3,11 @@ package com.zhuinden.simplestackextensionsample.features.registration
 import androidx.fragment.app.Fragment
 import com.zhuinden.simplestack.ScopeKey
 import com.zhuinden.simplestackextensions.fragments.DefaultFragmentKey
-import kotlinx.android.parcel.Parcelize
+import kotlinx.parcelize.Parcelize
 
 @Parcelize
-data class EnterProfileDataKey(private val placeholder: String = "") : DefaultFragmentKey(), ScopeKey.Child {
+data object EnterProfileDataKey : DefaultFragmentKey(), ScopeKey.Child {
     override fun instantiateFragment(): Fragment = EnterProfileDataFragment()
 
-    override fun getParentScopes(): List<String> = listOf("registration")
+    override fun getParentScopes(): List<String> = listOf(RegistrationViewModel::class.java.name)
 }